test(runtime): stabilize maintenance validation
